add: tally txs per block spending ephemeral dust
Eclair and LDK have recently reached cross-compatibility on zero-fee
commitment channels, and LDK has just released v0.2 which allows people
to open such channels.

Nonetheless, the commitment transactions of zero-fee commitment channels
will propagate to miners only if their nodes accept ephemeral dust.
Timely confirmation of commitment transactions is crucial to the
security model of Lightning, hence the need to observe which mining
pools mine ephemeral dust.

Here, we make a best-effort attempt at only counting the transactions in
a block spending ephemeral dust that could have been submitted via the
P2P network. We are not interested in coinbase transactions, or
transactions submitted out-of-band. Therefore, we assume that any
transaction spending ephemeral dust is version 3, non-zero-fee, smaller
than 1000vB, and spends from a version 3, zero-fee transaction itself
smaller than 10_000vB. See TRUC policy rules defined in BIP-431 for
further details.

For now, we do not assert that the parent transaction does not have any
ancestors in the same block, or that the child transaction does not have
any descendants in the same block (this would violate current policy on
TRUC packages). This check could be added in the future should
false-positives appear in the data.